Add support for inverted index tree:
> The system described in this paper provides keyword search functionality for a
DHT-based file system or archival storage system, to map keyword queries to the unique
routing keys described above. It does so by mapping each keyword to a node in the DHT
that will store a list of documents containing that keyword.

Now hash table, which is now an inverted index tree, is represented as follows:
"keyword1" ->
    "filename1", "ip1"
    "filename2", "ip2"
"keyword2" ->
    "filename3", "ip3"
  ...

Make changes in proto services and definitions to be able to Put and Get file records (filename and owner's IP), following changes in implementations of Put() and Get() in chordServerImpl.go, following changes in stubPut() and stubGet().

Change hash table to adapt to store file records as values in key-value pairs instead of simple strings.

Change PutInDHT() and FindInDHT() to be able to put a file record in DHT and get file records by a keyword.

Adapt ChordNode.String() to better print new hash table.